Extract the Blockfrost adapter so both providers share one schema.
Move Koios-shaped Blockfrost mapping out of util.js into a typed
providers module, and cover the mappers plus request fall-through.
